+++ src/logger/constants.ts
@@ -1,0 +1,5 @@
+export const LOCAL_STORAGE_KEY = "logs";
+export const INDEXED_DB_KEY = "logs";
+export const LOG_VERSION = 1;
+export const LOG_FILE_NAME = "log";
+export const LOG_FILE_FORMAT = "json";
