+++ src/logger/index.ts
@@ -1,2 +1,1 @@
 export { default as logger } from "./logger";
-export { downloadLogsAsJson } from "./utils";
